feat(localFileSearch): implement local file search module with configuration and results display
- Add LocalFileSearchTab as the main entry point for local file search functionality.
- Create LocalFileSearchConfigPanel for configuring search parameters such as result limit, match scope, and file extensions.
- Implement LocalFileSearchHeader to display the search title and result count.
- Develop LocalFileSearchQueryRow for inputting search keywords and triggering searches.
- Add LocalFileSearchResults to show search results with loading and empty states.
- Introduce LocalFileSearchRootRow for selecting the root directory for searches.
- Create utility functions for parsing CSV values for file extensions and excluded directories.
- Implement a custom hook (useLocalFileSearch) to manage the state and logic of the local file search.
- Define TypeScript types for better type safety and clarity in component props and hook return values.
- Add configuration constants for persistent storage of the search root directory.
