+++ src/renderer/components/states/maxExpand/components/localFileSearch/components/LocalFileSearchQueryRow.tsx
@@ -1,0 +1,67 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file LocalFileSearchQueryRow.tsx
+ * @description 搜索关键词输入行组件。
+ * @author 鸡哥
+ */
+
+import type { ReactElement } from 'react';
+import { useTranslation } from 'react-i18next';
+import type { LocalFileSearchQueryRowProps } from '../types/localFileSearchTypes';
+
+/**
+ * 搜索关键词输入行
+ * @description 输入关键词、搜索按钮、配置切换按钮
+ */
+export function LocalFileSearchQueryRow({
+  keyword, setKeyword, loading, showConfig, setShowConfig, onSearch,
+}: LocalFileSearchQueryRowProps): ReactElement {
+  const { t } = useTranslation();
+
+  return (
+    <div className="local-file-search-query-row">
+      <input
+        className="local-file-search-query-input"
+        value={keyword}
+        onChange={(e) => setKeyword(e.target.value)}
+        onKeyDown={(e) => {
+          if (e.key === 'Enter') {
+            onSearch();
+          }
+        }}
+        placeholder={t('maxExpand.localFileSearch.keywordPlaceholder', { defaultValue: '输入文件名关键词，例如 report、.pdf' })}
+      />
+      <button type="button" className="local-file-search-btn" onClick={onSearch}>
+        {loading
+          ? t('maxExpand.localFileSearch.searching', { defaultValue: '搜索中…' })
+          : t('maxExpand.localFileSearch.search', { defaultValue: '搜索' })}
+      </button>
+      <button
+        type="button"
+        className={`local-file-search-btn local-file-search-config-toggle${showConfig ? ' active' : ''}`}
+        onClick={() => setShowConfig((prev) => !prev)}
+      >
+        {t('maxExpand.localFileSearch.config', { defaultValue: '配置' })}
+      </button>
+    </div>
+  );
+}
